requirements.txt
alembic==1.8.1
Beaker==1.11.0
boto==2.49.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.1
defusedxml==0.7.1
dnspython==2.2.1
Flask==2.2.2
Flask-Migrate==3.1.0
Flask-PyMongo==2.3.0
Flask-pyoidc==3.11.0
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
future==0.18.2
greenlet==1.1.3.post0
gunicorn==21.2.0
idna==3.4
importlib-metadata==5.0.0
importlib-resources==5.10.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.3
MarkupSafe==2.1.1
oic==1.4.0
packaging==23.2
psycopg2==2.9.4
pycparser==2.21
pycryptodomex==3.15.0
pyjwkest==1.4.2
pymongo==4.3.2
PyMySQL==1.0.2
pytz==2022.5
radosgw-admin==1.7.2
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.42
SQLAlchemy-Utils==0.38.3
typing_extensions==4.4.0
urllib3==1.26.12
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.9.0
